--- a/staging/src/k8s.io/api/testdata/HEAD/core.v1.CreateOptions.yaml
+++ b/staging/src/k8s.io/api/testdata/HEAD/core.v1.CreateOptions.yaml
@@ -2,4 +2,5 @@ apiVersion: v1
 dryRun:
 - "2"
 fieldManager: "3"
+fieldValidation: "4"
 kind: CreateOptions