--- a/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.proto
+++ b/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.proto
@@ -242,6 +242,19 @@ message CreateOptions {
   // as defined by https://golang.org/pkg/unicode/#IsPrint.
   // +optional
   optional string fieldManager = 3;
+
+  // fieldValidation determines how the server should respond to
+  // unknown/duplicate fields in the object in the request.
+  // Introduced as alpha in 1.23, older servers or servers with the
+  // `ServerSideFieldValidation` feature disabled will discard valid values
+  // specified in  this param and not perform any server side field validation.
+  // Valid values are:
+  // - Ignore: ignores unknown/duplicate fields.
+  // - Warn: responds with a warning for each
+  // unknown/duplicate field, but successfully serves the request.
+  // - Strict: fails the request on unknown/duplicate fields.
+  // +optional
+  optional string fieldValidation = 4;
 }
 
 // DeleteOptions may be provided when deleting an API object.
@@ -878,6 +891,19 @@ message PatchOptions {
   // types (JsonPatch, MergePatch, StrategicMergePatch).
   // +optional
   optional string fieldManager = 3;
+
+  // fieldValidation determines how the server should respond to
+  // unknown/duplicate fields in the object in the request.
+  // Introduced as alpha in 1.23, older servers or servers with the
+  // `ServerSideFieldValidation` feature disabled will discard valid values
+  // specified in  this param and not perform any server side field validation.
+  // Valid values are:
+  // - Ignore: ignores unknown/duplicate fields.
+  // - Warn: responds with a warning for each
+  // unknown/duplicate field, but successfully serves the request.
+  // - Strict: fails the request on unknown/duplicate fields.
+  // +optional
+  optional string fieldValidation = 4;
 }
 
 // Preconditions must be fulfilled before an operation (update, delete, etc.) is carried out.
@@ -1095,6 +1121,19 @@ message UpdateOptions {
   // as defined by https://golang.org/pkg/unicode/#IsPrint.
   // +optional
   optional string fieldManager = 2;
+
+  // fieldValidation determines how the server should respond to
+  // unknown/duplicate fields in the object in the request.
+  // Introduced as alpha in 1.23, older servers or servers with the
+  // `ServerSideFieldValidation` feature disabled will discard valid values
+  // specified in  this param and not perform any server side field validation.
+  // Valid values are:
+  // - Ignore: ignores unknown/duplicate fields.
+  // - Warn: responds with a warning for each
+  // unknown/duplicate field, but successfully serves the request.
+  // - Strict: fails the request on unknown/duplicate fields.
+  // +optional
+  optional string fieldValidation = 3;
 }
 
 // Verbs masks the value so protobuf can generate