--- a/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/zz_generated.conversion.go
+++ b/staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/zz_generated.conversion.go
@@ -294,6 +294,13 @@ func autoConvert_url_Values_To_v1_CreateOptions(in *url.Values, out *CreateOptio
 	} else {
 		out.FieldManager = ""
 	}
+	if values, ok := map[string][]string(*in)["fieldValidation"]; ok && len(values) > 0 {
+		if err := runtime.Convert_Slice_string_To_string(&values, &out.FieldValidation, s); err != nil {
+			return err
+		}
+	} else {
+		out.FieldValidation = ""
+	}
 	return nil
 }
 
@@ -449,6 +456,13 @@ func autoConvert_url_Values_To_v1_PatchOptions(in *url.Values, out *PatchOptions
 	} else {
 		out.FieldManager = ""
 	}
+	if values, ok := map[string][]string(*in)["fieldValidation"]; ok && len(values) > 0 {
+		if err := runtime.Convert_Slice_string_To_string(&values, &out.FieldValidation, s); err != nil {
+			return err
+		}
+	} else {
+		out.FieldValidation = ""
+	}
 	return nil
 }
 
@@ -496,6 +510,13 @@ func autoConvert_url_Values_To_v1_UpdateOptions(in *url.Values, out *UpdateOptio
 		}
 	} else {
 		out.FieldManager = ""
+	}
+	if values, ok := map[string][]string(*in)["fieldValidation"]; ok && len(values) > 0 {
+		if err := runtime.Convert_Slice_string_To_string(&values, &out.FieldValidation, s); err != nil {
+			return err
+		}
+	} else {
+		out.FieldValidation = ""
 	}
 	return nil
 }