… api/openapi-spec/v3/api__v1_openapi.json
… api/openapi-spec/v3/apis__admissionregistration.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__apiextensions.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__apps__v1_openapi.json
… api/openapi-spec/v3/apis__authentication.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__authorization.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__autoscaling__v1_openapi.json
… api/openapi-spec/v3/apis__autoscaling__v2_openapi.json
… api/openapi-spec/v3/apis__autoscaling__v2beta1_openapi.json
… api/openapi-spec/v3/apis__autoscaling__v2beta2_openapi.json
… api/openapi-spec/v3/apis__batch__v1_openapi.json
… api/openapi-spec/v3/apis__batch__v1beta1_openapi.json
… api/openapi-spec/v3/apis__certificates.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__coordination.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__discovery.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__discovery.k8s.io__v1beta1_openapi.json
… api/openapi-spec/v3/apis__events.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__events.k8s.io__v1beta1_openapi.json
… api/openapi-spec/v3/apis__flowcontrol.apiserver.k8s.io__v1beta1_openapi.json
… api/openapi-spec/v3/apis__flowcontrol.apiserver.k8s.io__v1beta2_openapi.json
… api/openapi-spec/v3/apis__internal.apiserver.k8s.io__v1alpha1_openapi.json
… api/openapi-spec/v3/apis__networking.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__node.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__node.k8s.io__v1alpha1_openapi.json
… api/openapi-spec/v3/apis__node.k8s.io__v1beta1_openapi.json
… api/openapi-spec/v3/apis__policy__v1_openapi.json
… api/openapi-spec/v3/apis__policy__v1beta1_openapi.json
… api/openapi-spec/v3/apis__rbac.authorization.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__scheduling.k8s.io__v1_openapi.json
… api/openapi-spec/v3/apis__storage.k8s.io__v1_openapi.json