@kevindelgado 2021-11-11
YAML parsing fixed. All feedback should be addressed now.

PTAL @liggitt @apelisse