@kevindelgado 2021-11-12
All feedback addressed. All tests passing. Ready for another review.

@liggitt @apelisse