@SergeyKanzhelev 2021-11-15
/remove-sig node