@kevindelgado 2021-11-16
Fixed all the nagging CI issues that my last commit broke. Everything is good to go for another review @liggitt 

Spoke to @apelisse about the failing gce-performance tests and those are unrelated to the code.