@kevindelgado 2021-11-18
benchmark results:

### old way vs new way (Ignore)
```
name                                      old time/op    new time/op    delta
FromUnstructuredWithValidation/Ignore-16     149µs ± 1%     160µs ± 1%  +7.01%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
FromUnstructuredWithValidation/Ignore-16    26.8kB ± 0%    29.5kB ± 0%  +9.99%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
FromUnstructuredWithValidation/Ignore-16       641 ± 0%       663 ± 0%  +3.43%  (p=0.008 n=5+5)
```

### old way vs new way (Strict)
```
name                                      old time/op    new time/op    delta
FromUnstructuredWithValidation/Strict-16     149µs ± 1%     169µs ± 1%  +13.43%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
FromUnstructuredWithValidation/Strict-16    26.8kB ± 0%    32.5kB ± 0%  +21.20%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
FromUnstructuredWithValidation/Strict-16       641 ± 0%       732 ± 0%  +14.20%  (p=0.008 n=5+5)
```