@kevindelgado 2021-11-18
Benchmark added. Verified that new logic is _mostly_ inert for the Ignore case.

PTAL @apelisse @liggitt